Create buttons after reading XML document
After reading the downloaded XML document, create buttons representing
every APK that the user should be able to download. These buttons are
created in the order they are read in the XML document, are alligned in
the center and are always placed "under" the above button.
This means that when the Cancel Download button appear, all other
buttons are moved down a bit.
The button does not do anything else at this moment other than create
the Cancel Download button. Pressing that button will crash the code.