Nearing the end! (request input if code is invalid, if download failed and if file does not exist)
The code is starting to look completed, though we are not there just
yet. Some error checks and feedback to the user remains.

In the beginning of the code we still check for required persmission,
and prompt the user to allow them if they have not yet been added.
After that, the program looks for the save file. If found, it will read
the code. It then tries to download a file, using the text in the file
as part of the download path. If it fails, the user will be prompted to
write the code anew.

If the file is not found, the program will promt the user to write the
code. It will then try to download and save and whatnot again from
there.