Shows a toast after X seconds, even if the app has been closed!
I'm going to attempt to user clock-timer to check for APK updates every
24-hours.
Whenever the user closes the app it will initiate a new "wait for 24
hours" thingy, and after 24 hours the app will download the XML file,
check the version number of the XML document and the version saved on
the phone.
If the versions are different, at least one of the apps has been
updated! This will send a notification to the user explaining that "One
or more apps requires an update.".

Whenever the user enters the app, the app will stop the current "wait
for 24 hours" whatever, and again check for actual updates.