Added new project that is the start of a widget! This is to be used for the Hurtigruta case.
We want to have a widget on the home screen, which will have a button to
launch the Hurtigruta app (dunno what sort of app this is yet, info yet
to come.), and a button to search for app updates.

This widget will search for updates for both itself and the other app,
giving the user a super smooth experience.