Movies and posters are downloaded!
They are downloaded into the temp folders, which will be deleted if the
user either delete the app, or cleans the app's cache or data. But, as
soon as either a poster or movie is downloaded the program will move
them to the Movie folder.

Now to figure out why the app always re-downloads the apk files.